+++ include/UDRefl/FieldPtr.h
@@ -79,6 +79,7 @@ namespace Ubpa::UDRefl {
 		constexpr bool IsStatic()        const noexcept { return data.index() == 2; }
 		constexpr bool IsDynamicShared() const noexcept { return data.index() == 3; }
 		constexpr bool IsDyanmicBuffer() const noexcept { return data.index() == 4; }
+		constexpr FieldFlag GetFieldFlag() const noexcept;
 
 		constexpr bool IsOwned()   const noexcept { return data.index() <  2; }
 		constexpr bool IsUnowned() const noexcept { return data.index() >= 2; }
@@ -94,3 +95,5 @@ namespace Ubpa::UDRefl {
 		Data data;
 	};
 }
+
+#include "details/FieldPtr.inl"