+++ include/UDRefl/Object.h
@@ -97,7 +97,7 @@ namespace Ubpa::UDRefl {
 		// Invoke
 		///////////
 
-		InvocableResult IsInvocable(Name method_name, std::span<const Type> argTypes = {}, FuncFlag mode = FuncFlag::All) const;
+		InvocableResult IsInvocable(Name method_name, std::span<const Type> argTypes = {}, FuncFlag flag = FuncFlag::All) const;
 
 		InvokeResult Invoke(
 			Name method_name,
@@ -107,7 +107,7 @@ namespace Ubpa::UDRefl {
 			FuncFlag flag = FuncFlag::All) const;
 
 		template<typename... Args>
-		InvocableResult IsInvocable(Name method_name, FuncFlag mode = FuncFlag::All) const;
+		InvocableResult IsInvocable(Name method_name, FuncFlag flag = FuncFlag::All) const;
 
 		template<typename T>
 		T InvokeRet(Name method_name, std::span<const Type> argTypes = {}, ArgPtrBuffer argptr_buffer = nullptr, FuncFlag flag = FuncFlag::All) const;
@@ -165,11 +165,9 @@ namespace Ubpa::UDRefl {
 		// Fields
 		///////////
 
-		// all
-		ObjectView Var(Name field_name) const;
-
-		// all, for diamond inheritance
-		ObjectView Var(Type base, Name field_name) const;
+		ObjectView Var(Name field_name, FieldFlag flag = FieldFlag::All) const;
+		// for diamond inheritance
+		ObjectView Var(Type base, Name field_name, FieldFlag flag = FieldFlag::All) const;
 
 		// self vars and all bases' vars
 		void ForEachVar(const std::function<bool(TypeRef, FieldRef, ObjectView)>& func) const;
@@ -200,7 +198,7 @@ namespace Ubpa::UDRefl {
 		bool ContainsBase  (Type base       ) const;
 		bool ContainsField (Name field_name ) const;
 		bool ContainsMethod(Name method_name) const;
-		bool ContainsMethod(Name method_name, FuncFlag mode) const;
+		bool ContainsMethod(Name method_name, FuncFlag flag) const;
 
 		//
 		// Type