+++ src/core/ReflMngr.cpp
@@ -1156,62 +1156,27 @@ ObjectView ReflMngr::DynamicCast(ObjectView obj, Type type) const {
 	return nullptr;
 }
 
-ObjectView ReflMngr::Var(Type type, Name field_name) {
-	const CVRefMode cvref_mode = type.CVRefMode();
-	assert(!CVRefMode_IsVolatile(cvref_mode));
-	switch (cvref_mode)
-	{
-	case CVRefMode::Left:
-		return Var(type.RemoveReference(), field_name).AddLValueReference();
-	case CVRefMode::Right:
-		return Var(type.RemoveReference(), field_name).AddRValueReference();
-	case CVRefMode::Const:
-		return Var(type.RemoveConst(), field_name).AddConst();
-	case CVRefMode::ConstLeft:
-		return Var(type.RemoveCVRef(), field_name).AddConstLValueReference();
-	case CVRefMode::ConstRight:
-		return Var(type.RemoveCVRef(), field_name).AddConstRValueReference();
-	default:
-		break;
-	}
-
-	auto ttarget = typeinfos.find(type);
-	if (ttarget == typeinfos.end())
-		return nullptr;
-
-	auto& typeinfo = ttarget->second;
-
-	auto ftarget = typeinfo.fieldinfos.find(field_name);
-	if (ftarget != typeinfo.fieldinfos.end() && ftarget->second.fieldptr.IsUnowned())
-		return ftarget->second.fieldptr.Var();
-
-	for (const auto& [base, baseinfo] : typeinfo.baseinfos) {
-		auto bptr = Var(base, field_name);
-		if (bptr.GetType())
-			return bptr;
-	}
-
-	return nullptr;
-}
-
-ObjectView ReflMngr::Var(ObjectView obj, Name field_name) {
+ObjectView ReflMngr::Var(ObjectView obj, Name field_name, FieldFlag flag) {
 	const CVRefMode cvref_mode = obj.GetType().CVRefMode();
 	assert(!CVRefMode_IsVolatile(cvref_mode));
 	switch (cvref_mode)
 	{
 	case CVRefMode::Left:
-		return Var(obj.RemoveReference(), field_name).AddLValueReference();
+		return Var(obj.RemoveReference(), field_name, flag).AddLValueReference();
 	case CVRefMode::Right:
-		return Var(obj.RemoveReference(), field_name).AddRValueReference();
+		return Var(obj.RemoveReference(), field_name, flag).AddRValueReference();
 	case CVRefMode::Const:
-		return Var(obj.RemoveConst(), field_name).AddConst();
+		return Var(obj.RemoveConst(), field_name, flag).AddConst();
 	case CVRefMode::ConstLeft:
-		return Var(obj.RemoveConstReference(), field_name).AddConstLValueReference();
+		return Var(obj.RemoveConstReference(), field_name, flag).AddConstLValueReference();
 	case CVRefMode::ConstRight:
-		return Var(obj.RemoveConstReference(), field_name).AddConstRValueReference();
+		return Var(obj.RemoveConstReference(), field_name, flag).AddConstRValueReference();
 	default:
 		break;
 	}
+
+	if (!obj.GetPtr())
+		flag = enum_within(flag, FieldFlag::Unowned);
 
 	auto ttarget = typeinfos.find(obj.GetType());
 	if (ttarget == typeinfos.end())
@@ -1220,7 +1185,7 @@ ObjectView ReflMngr::Var(ObjectView obj, Name field_name) {
 	auto& typeinfo = ttarget->second;
 
 	auto ftarget = typeinfo.fieldinfos.find(field_name);
-	if (ftarget != typeinfo.fieldinfos.end())
+	if (ftarget != typeinfo.fieldinfos.end() && enum_contain(flag, ftarget->second.fieldptr.GetFieldFlag()))
 		return ftarget->second.fieldptr.Var(obj.GetPtr());
 
 	for (const auto& [base, baseinfo] : typeinfo.baseinfos) {
@@ -1232,7 +1197,7 @@ ObjectView ReflMngr::Var(ObjectView obj, Name field_name) {
 	return nullptr;
 }
 
-ObjectView ReflMngr::Var(ObjectView obj, Type base, Name field_name) {
+ObjectView ReflMngr::Var(ObjectView obj, Type base, Name field_name, FieldFlag flag) {
 	auto base_obj = StaticCast_DerivedToBase(obj, base);
 	if (!base_obj.GetType())
 		return nullptr;
@@ -1365,7 +1330,7 @@ InvokeResult ReflMngr::Invoke(
 	}
 
 	if (!obj.GetPtr())
-		flag = enum_remove(flag, FuncFlag::Variable | FuncFlag::Const);
+		flag = enum_within(flag, FuncFlag::Static);
 
 	if (auto priority_rst = details::Invoke(true, &temporary_resource, rawObj, method_name, result_buffer, argTypes, argptr_buffer, flag))
 		return priority_rst;
@@ -1407,7 +1372,7 @@ SharedObject ReflMngr::MInvoke(
 	}
 
 	if (!obj.GetPtr())
-		flag = enum_remove(flag, FuncFlag::Variable | FuncFlag::Const);
+		flag = enum_within(flag, FuncFlag::Static);
 
 	if (auto priority_rst = details::MInvoke(true, &temporary_resource, rawObj, method_name, rst_rsrc, argTypes, argptr_buffer, flag); priority_rst.GetType().Valid())
 		return priority_rst;
@@ -1963,7 +1928,7 @@ bool ReflMngr::ContainsMethod(Type type, Name method_name) const {
 	return false;
 }
 
-bool ReflMngr::ContainsMethod(Type type, Name method_name, FuncFlag mode) const {
+bool ReflMngr::ContainsMethod(Type type, Name method_name, FuncFlag flag) const {
 	auto target = typeinfos.find(type);
 	if (target == typeinfos.end())
 		return false;
@@ -1971,12 +1936,12 @@ bool ReflMngr::ContainsMethod(Type type, Name method_name, FuncFlag mode) const 
 	const auto& info = target->second;
 	auto [begin_iter, end_iter] = info.methodinfos.equal_range(method_name);
 	for (auto iter = begin_iter; iter != end_iter; ++iter) {
-		if (iter->second.methodptr.GetFuncFlag() == mode)
+		if (iter->second.methodptr.GetFuncFlag() == flag)
 			return true;
 	}
 
 	for (const auto& [basetype, baseinfo] : info.baseinfos) {
-		if (ContainsMethod(basetype, method_name, mode))
+		if (ContainsMethod(basetype, method_name, flag))
 			return true;
 	}